Fixed date range event after a range was fixed
When the user clicks a end date > start date or start date > end date,
the other date is automatically set to the same date.
Now the "startDateSelected" and "endDateSelected" events are now sending
updated dates.

Also, the extra "onDateSelected" event is prevented for the date that's
getting fixed, as this one is passed through by the other events
already.