Remove timepicker support when `singleDate: true`
When using `singleDate: true` the daterangepicker renders only
one calendar. This breaks the timepicker, which is not able
to format a date range string when the end calendar is missing.
Support was apparently partially implemented, but since it was missing
on the demo page it was never noticed as being broken.
This commit removes this partial support and makes the datepicker
throw when instantiated with `singleDate: true` _and_ the timepicker
plugin. This is now tested.