Add timezone support to the daterangepicker
The daterangepicker now uses the `timezone` parameter passed to it
not just to display it next to the timepickers, but also to parse
the <year, month, day, (hour, minutes)?> tuple.
This means that the dateFormatter is now passed start, end dates
which are the result of parsing the displayed (start, end) datetimes
in the timezone passed as parameter.
If a timezone parameter is not supplied, all parsing is done in UTC.